fix(mas,mesh,fft): honour float64 end-to-end on CPU and fix MAS box-edge handling
Precision (type-neutrality):
- mas.assign/readout no longer force float32 on the CPU; the grid follows the
  mesh's working precision and positions/weights are cast to match, so a
  dtype=float64 run is honoured during painting instead of being silently
  downcast. GPU stays float32 by design.
- Mesh stores boxsize/boxcentre (and derived cell_size/min_corner) at the mesh
  dtype instead of pinning to float32; float32 mode is unchanged.
- The iFFT/Burden CPU solver honours the dtype: scipy and pyfftw wavevectors and
  the pyfftw in-place buffers/outputs follow the delta dtype (f32/c64 or
  f64/c128). prepare_k_components now also accepts a numpy.dtype instance.

MAS non-periodic boundary correctness:
- tsc_assign_serial ignored its pbc argument (always wrapped); it now honours
  pbc, clamping to the boundary when pbc=False.
- pbc=False now clamps (mass-conserving) uniformly across NGP/CIC/TSC and the
  serial and parallel kernels; the parallel kernels and tsc_read previously
  dropped out-of-range contributions. Paint/read are now mutual adjoints at the
  edge, and serial==parallel.
- cic_assign_serial/cic_read use floor instead of int() truncation (correct for
  negative positions under PBC).

Bump version to 0.5.0; document in CHANGELOG and correct the pyfftw doc's
float32-only output claim.

+++ baorecon/solvers/fft/cpu.py
@@ -54,7 +54,11 @@ class FFTSolverCPU(PoissonSolver):
         pure waste).
         """
         if self._k_host is None:
-            self._k_host = prepare_k_components(self.mesh.cell_size, self.mesh.nmesh)
+            # Build the wavevectors at the mesh's working precision so a float64
+            # mesh yields float64 k (and hence a float64 potential/displacement).
+            self._k_host = prepare_k_components(
+                self.mesh.cell_size, self.mesh.nmesh, dtype=self.mesh.dtype
+            )
         return self._k_host
 
     def _compute_displacement_mesh(self) -> None:
